+++ src/solvers/cheese.rs
@@ -115,7 +115,7 @@ fn cheese_solver() {
     eprintln!("cost: {}", painter.cost);
     let dist = image_distance(&img, &target);
     eprintln!("distance to target: {}", dist);
-    eprintln!("final score: {}", dist.round() as i32 + painter.cost);
+    eprintln!("final score: {}", dist.round() as i64 + painter.cost);
 
     let output_path = format!("outputs/cheese_{}.png", problem_id);
     img.save(&crate::util::project_path(&output_path));
